Adding new paragraph to Concrete
added
It is important to discuss the nature of your claim with an attorney.  The requirements of the law are fact-specific, often requiring an underlying condition. Understanding the claim, and explaining the requirements to your doctor may be necessary. Contact my office to ensure you get the care and compensation you deserve.